Move gatt handling code from l2cap to conn
pure code movement
